requirements.txt
python-dotenv==1.1.1
langchain==0.3.26
langchain-openai==0.3.27
qdrant-client==1.15.0
streamlit==1.48.1